move tests for assessment-action-message-creator to separate test file
